Making requested changes to response object
* Replacing simple string message with a more detailed JSON response, breaking
down the number of success and failures as well as returning a list of codes and
messages for each submission changed
* The structuring of the response is now handled by a
`__prepare_bulk_update_response` method
* An additional check was added for the length of `submission_ids` in the
payload. Previously, if it was an empty array, all submissions would be edited.
Now a validation error is sent to the client, requiring at least one value to
the in the array